+++ atlas/src/components/ErrorFallback.tsx
@@ -1,48 +1,30 @@
 import React from 'react'
+
 import styled from '@emotion/styled'
 import { FallbackProps } from 'react-error-boundary'
 
-import { ReactComponent as ErrorIllustration } from '@/assets/error.svg'
-import { Button, Typography } from '@/shared/components'
-import { sizes } from '@/shared/theme'
+import { Button } from '@/shared/components'
+import { sizes, colors } from '@/shared/theme'
 
 const Container = styled.div`
-  margin: ${sizes.b10 * 2}px auto 0;
+  padding: ${sizes.b4};
+  color: ${colors.gray[400]};
   display: grid;
   place-items: center;
-
-  > svg {
-    max-width: 650px;
-  }
 `
 
-const Message = styled.div`
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  margin-top: 90px;
-  margin-bottom: ${sizes.b10}px;
-
-  > p {
-    margin: 0;
-    line-height: 1.75;
-  }
-`
-
-const Title = styled(Typography)`
-  line-height: 1.25;
+const StyledButton = styled(Button)`
+  color: ${colors.white};
 `
 
 const ErrorFallback: React.FC<FallbackProps> = ({ error, resetErrorBoundary }) => {
   console.error(error)
   return (
     <Container>
-      <ErrorIllustration />
-      <Message>
-        <Title variant="h3">Ops! An Error occurred</Title>
-        <p>We could not aquire expected results. Please try reloading or return to the home page.</p>
-      </Message>
-      <Button onClick={resetErrorBoundary}>Return to home page</Button>
+      <p>Something went wrong...</p>
+      <StyledButton variant="tertiary" onClick={resetErrorBoundary}>
+        Try again
+      </StyledButton>
     </Container>
   )
 }
