+++ atlas/src/components/index.ts
@@ -8,5 +8,5 @@ export { default as PlaceholderVideoGrid } from './PlaceholderVideoGrid'
 export { default as VideoPreview } from './VideoPreviewWithNavigation'
 export { default as ChannelPreview } from './ChannelPreviewWithNavigation'
 export { default as ChannelGrid } from './ChannelGrid'
+export { default as ViewErrorFallback } from './ViewErrorFallback'
 export { default as ErrorFallback } from './ErrorFallback'
-export { default as GalleryErrorFallback } from './GalleryErrorFallback'
